+++ src/platform_impl/macos/monitor.rs
@@ -6,19 +6,21 @@ use core_foundation::{
     array::{CFArrayGetCount, CFArrayGetValueAtIndex},
     base::{CFRelease, TCFType},
     string::CFString,
+    uuid::CFUUIDRef,
 };
-use core_graphics::display::{CGDirectDisplayID, CGDisplay, CGDisplayBounds};
+use core_graphics::display::{CGDirectDisplayID, CGDisplay};
 use objc2::rc::{Id, Shared};
 
 use super::appkit::NSScreen;
 use super::ffi;
-use crate::dpi::{PhysicalPosition, PhysicalSize};
+use crate::dpi::{LogicalPosition, LogicalSize, PhysicalPosition, PhysicalSize};
+use crate::monitor::MonitorGone;
 
 #[derive(Clone)]
 pub struct VideoMode {
-    pub(crate) size: (u32, u32),
-    pub(crate) bit_depth: u16,
-    pub(crate) refresh_rate_millihertz: u32,
+    size: PhysicalSize<u32>,
+    bit_depth: u16,
+    refresh_rate_millihertz: u32,
     pub(crate) monitor: MonitorHandle,
     pub(crate) native_mode: NativeDisplayMode,
 }
@@ -77,7 +79,7 @@ impl Clone for NativeDisplayMode {
 
 impl VideoMode {
     pub fn size(&self) -> PhysicalSize<u32> {
-        self.size.into()
+        self.size
     }
 
     pub fn bit_depth(&self) -> u16 {
@@ -101,10 +103,7 @@ pub struct MonitorHandle(CGDirectDisplayID);
 // unique identifier that persists even across system reboots
 impl PartialEq for MonitorHandle {
     fn eq(&self, other: &Self) -> bool {
-        unsafe {
-            ffi::CGDisplayCreateUUIDFromDisplayID(self.0)
-                == ffi::CGDisplayCreateUUIDFromDisplayID(other.0)
-        }
+        self.uuid() == other.uuid()
     }
 }
 
@@ -118,18 +117,13 @@ impl PartialOrd for MonitorHandle {
 
 impl Ord for MonitorHandle {
     fn cmp(&self, other: &Self) -> std::cmp::Ordering {
-        unsafe {
-            ffi::CGDisplayCreateUUIDFromDisplayID(self.0)
-                .cmp(&ffi::CGDisplayCreateUUIDFromDisplayID(other.0))
-        }
+        self.uuid().cmp(&other.uuid())
     }
 }
 
 impl std::hash::Hash for MonitorHandle {
     fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
-        unsafe {
-            ffi::CGDisplayCreateUUIDFromDisplayID(self.0).hash(state);
-        }
+        self.uuid().hash(state);
     }
 }
 
@@ -151,26 +145,15 @@ pub fn primary_monitor() -> MonitorHandle {
 
 impl fmt::Debug for MonitorHandle {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        // TODO: Do this using the proper fmt API
-        #[derive(Debug)]
-        #[allow(dead_code)]
-        struct MonitorHandle {
-            name: Option<String>,
-            native_identifier: u32,
-            size: PhysicalSize<u32>,
-            position: PhysicalPosition<i32>,
-            scale_factor: f64,
-        }
-
-        let monitor_id_proxy = MonitorHandle {
-            name: self.name(),
-            native_identifier: self.native_identifier(),
-            size: self.size(),
-            position: self.position(),
-            scale_factor: self.scale_factor(),
-        };
-
-        monitor_id_proxy.fmt(f)
+        f.debug_struct("MonitorHandle")
+            .field("name", &self.name())
+            .field("native_identifier", &self.native_identifier())
+            .field("size", &self.size())
+            .field("position", &self.position())
+            .field("scale_factor", &self.scale_factor())
+            .field("refresh_rate_millihertz", &self.refresh_rate_millihertz())
+            .field("uuid", &self.uuid())
+            .finish_non_exhaustive()
     }
 }
 
@@ -179,10 +162,16 @@ impl MonitorHandle {
         MonitorHandle(id)
     }
 
-    pub fn name(&self) -> Option<String> {
+    // TODO: Be smarter about this:
+    // <https://github.com/glfw/glfw/blob/57cbded0760a50b9039ee0cb3f3c14f60145567c/src/cocoa_monitor.m#L44-L126>
+    pub fn name(&self) -> Result<String, MonitorGone> {
         let MonitorHandle(display_id) = *self;
         let screen_num = CGDisplay::new(display_id).model_number();
-        Some(format!("Monitor #{}", screen_num))
+        if screen_num == 0xFFFFFFFF {
+            Err(MonitorGone::new())
+        } else {
+            Ok(format!("Monitor #{}", screen_num))
+        }
     }
 
     #[inline]
@@ -190,58 +179,62 @@ impl MonitorHandle {
         self.0
     }
 
-    pub fn size(&self) -> PhysicalSize<u32> {
-        let MonitorHandle(display_id) = *self;
-        let display = CGDisplay::new(display_id);
-        let height = display.pixels_high();
-        let width = display.pixels_wide();
-        PhysicalSize::from_logical::<_, f64>((width as f64, height as f64), self.scale_factor())
-    }
-
-    #[inline]
-    pub fn position(&self) -> PhysicalPosition<i32> {
-        let bounds = unsafe { CGDisplayBounds(self.native_identifier()) };
-        PhysicalPosition::from_logical::<_, f64>(
-            (bounds.origin.x as f64, bounds.origin.y as f64),
-            self.scale_factor(),
-        )
-    }
-
-    pub fn scale_factor(&self) -> f64 {
-        match self.ns_screen() {
-            Some(screen) => screen.backingScaleFactor() as f64,
-            None => 1.0, // default to 1.0 when we can't find the screen
+    pub fn size(&self) -> Result<PhysicalSize<u32>, MonitorGone> {
+        // Weirdly enough, even though these methods are named "pixels",
+        // they're not actually physical pixels, but rather a logical unit.
+        let width = self.display().pixels_wide();
+        let height = self.display().pixels_high();
+        if !self.display().is_online() {
+            return Err(MonitorGone::new());
         }
+        Ok(LogicalSize::new(width as u32, height as u32)
+            .to_physical(self.scale_factor().unwrap_or(1.0)))
     }
 
-    pub fn refresh_rate_millihertz(&self) -> Option<u32> {
+    pub fn position(&self) -> Result<PhysicalPosition<i32>, MonitorGone> {
+        let bounds = self.display().bounds();
+        if !self.display().is_online() {
+            return Err(MonitorGone::new());
+        }
+        Ok(LogicalPosition::new(bounds.origin.x, bounds.origin.y)
+            .to_physical::<i32>(self.scale_factor().unwrap_or(1.0)))
+    }
+
+    pub fn scale_factor(&self) -> Result<f64, MonitorGone> {
+        self.ns_screen()
+            .map(|screen| screen.backingScaleFactor() as f64)
+    }
+
+    pub fn refresh_rate_millihertz(&self) -> Result<u32, MonitorGone> {
         unsafe {
             let mut display_link = std::ptr::null_mut();
             if ffi::CVDisplayLinkCreateWithCGDisplay(self.0, &mut display_link)
                 != ffi::kCVReturnSuccess
             {
-                return None;
+                return Err(MonitorGone::new());
             }
             let time = ffi::CVDisplayLinkGetNominalOutputVideoRefreshPeriod(display_link);
             ffi::CVDisplayLinkRelease(display_link);
 
             // This value is indefinite if an invalid display link was specified
             if time.flags & ffi::kCVTimeIsIndefinite != 0 {
-                return None;
+                return Err(MonitorGone::new());
             }
 
-            Some((time.time_scale as i64 / time.time_value * 1000) as u32)
+            Ok((time.time_scale as i64 / time.time_value * 1000) as u32)
         }
     }
 
-    pub fn video_modes(&self) -> impl Iterator<Item = VideoMode> {
-        let refresh_rate_millihertz = self.refresh_rate_millihertz().unwrap_or(0);
+    pub fn video_modes(&self) -> Result<impl Iterator<Item = VideoMode>, MonitorGone> {
+        let refresh_rate_millihertz = self.refresh_rate_millihertz()?;
         let monitor = self.clone();
 
         unsafe {
             let modes = {
                 let array = ffi::CGDisplayCopyAllDisplayModes(self.0, std::ptr::null());
-                assert!(!array.is_null(), "failed to get list of display modes");
+                if array.is_null() {
+                    return Err(MonitorGone::new());
+                }
                 let array_count = CFArrayGetCount(array);
                 let modes: Vec<_> = (0..array_count)
                     .map(move |i| {
@@ -254,14 +247,13 @@ impl MonitorHandle {
                 modes
             };
 
-            modes.into_iter().map(move |mode| {
-                let cg_refresh_rate_millihertz =
-                    ffi::CGDisplayModeGetRefreshRate(mode).round() as i64;
+            Ok(modes.into_iter().map(move |mode| {
+                let cg_refresh_rate_hertz = ffi::CGDisplayModeGetRefreshRate(mode);
 
                 // CGDisplayModeGetRefreshRate returns 0.0 for any display that
                 // isn't a CRT
-                let refresh_rate_millihertz = if cg_refresh_rate_millihertz > 0 {
-                    (cg_refresh_rate_millihertz * 1000) as u32
+                let refresh_rate_millihertz = if cg_refresh_rate_hertz > 0.0 {
+                    (cg_refresh_rate_hertz * 1000.0).round() as u32
                 } else {
                     refresh_rate_millihertz
                 };
@@ -280,7 +272,7 @@ impl MonitorHandle {
                 };
 
                 VideoMode {
-                    size: (
+                    size: PhysicalSize::new(
                         ffi::CGDisplayModeGetPixelWidth(mode) as u32,
                         ffi::CGDisplayModeGetPixelHeight(mode) as u32,
                     ),
@@ -289,23 +281,37 @@ impl MonitorHandle {
                     monitor: monitor.clone(),
                     native_mode: NativeDisplayMode(mode),
                 }
-            })
+            }))
         }
     }
 
-    pub(crate) fn ns_screen(&self) -> Option<Id<NSScreen, Shared>> {
-        let uuid = unsafe { ffi::CGDisplayCreateUUIDFromDisplayID(self.0) };
-        NSScreen::screens()
-            .into_iter()
-            .find(|screen| {
-                let other_native_id = screen.display_id();
-                let other_uuid = unsafe {
-                    ffi::CGDisplayCreateUUIDFromDisplayID(other_native_id as CGDirectDisplayID)
-                };
-                uuid == other_uuid
-            })
-            .map(|screen| unsafe {
-                Id::retain(screen as *const NSScreen as *mut NSScreen).unwrap()
-            })
+    pub(crate) fn ns_screen(&self) -> Result<Id<NSScreen, Shared>, MonitorGone> {
+        if let Some(uuid) = self.uuid() {
+            NSScreen::screens()
+                .into_iter()
+                .find(|screen| Self::new(screen.display_id()).uuid() == Some(uuid))
+                .map(|screen: &NSScreen| unsafe {
+                    Id::retain(screen as *const NSScreen as *mut NSScreen).unwrap()
+                })
+                .ok_or_else(|| {
+                    warn!("could not find NSScreen corresponding to monitor");
+                    MonitorGone::new()
+                })
+        } else {
+            Err(MonitorGone::new())
+        }
+    }
+
+    fn uuid(&self) -> Option<CFUUIDRef> {
+        let ptr = unsafe { ffi::CGDisplayCreateUUIDFromDisplayID(self.0) };
+        if ptr.is_null() {
+            None
+        } else {
+            Some(ptr)
+        }
+    }
+
+    fn display(&self) -> CGDisplay {
+        CGDisplay::new(self.0)
     }
 }